Remove all occurences of method=post
This hopefully fixes #8. The consequence is of course that the url will
include all search parameters, and so it won't look as clean. But it
also comes with the benefit of easier sharing of the exact site you're
on, which also is beneficial for us when debugging errors from users.