Implement prepare_script, for sanj. See full msg.
This commit implements the ability for a dictionary to declare a
field "prepare_script" in it's yaml config. See
configs/sanj.config.yaml.in for an example.

If a dictionary's config has a prepare_script, that script will be run _before_
`nds compile <proj>` tries to generate the final merged .xml file from
that dictionary's src/*.xml files. This way, dictionaries that has
their source material in a different format than the usual src/*.xml
files can now be built by `nds compile <proj>` by having a
prepare_script that takes care of building the src/*.xml file(s).

In the case of sanj, the sanj.config.yaml.in has been altered with such
a prepare_script. Sanj's prepare_script runs the script in the repo
that builds src/*.xml from inc/*.xslx.

Note: Sanjs prepare_script works as follows:
- find the gut root, to find the repo
- cd to the repo and activate the *already existing* virtual env (1)
- run the scripts/xslx2gtxml...py script to make src/<...>.xml from the
newest .xlsx file found in inc/

Note 2: A prepare_script will be given a PREPARE_SCRIPT_FORCE="yes"
environment variable IF `nds compile <proj>` was run with the `--force`
(or `-f` flag), as in `nds compile sanj --force`. The prepare_script in
sanj will handle this flag, and by default not rebuild the src/...xml
file if it's newer than the inc/...xlsx file. But, passing `--force` to
`nds compile ..` will make sure the prepare_script always runs.

The way this works is that the scripts in
`dict-(sjd-rus|rus-sjd)-x-private/scripts` has been given a new
`--skip=(no|if-newer)` argument, which defaults to `--skip=no` (always
build), but can be set by the prepare_script to `skip=if-newer` to skip
building src/...xml if it's newer than inc/...xslx.