Dont print analyses to stdout.
This change is done so as to not create needless lines in the log. This
is part of the process of getting logs to azure. As of right now, the
useful information is stored in morph_log, and in the nginx access log.
The former can be sent by nginx directy to the syslog, but the former
will have to be solved differently, through AMA log files, see
https://divvungiellatekno.github.io/giellalt.uit.no/infra/ServersLoggingToAzure.html